# docs/openapi.json
{
    "schemes": [],
    "swagger": "2.0",
    "info": {
        "description": "Task manager API",
        "title": "Task Manager API",
        "contact": {},
        "version": "dev"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/user/me": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns the user_id and email from the JWT claims.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "Get current user",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.MeEnvelope"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorEnvelope"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorEnvelope"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "dto.ErrorCode": {
            "type": "string",
            "enum": [
                "VALIDATION_ERROR",
                "INVALID_REQUEST",
                "UNAUTHORIZED",
                "FORBIDDEN",
                "PAGE_NOT_FOUND",
                "CONFLICT",
                "DATABASE_ERROR",
                "INTERNAL_ERROR",
                "INVALID_TOKEN",
                "MISSING_TOKEN",
                "INVALID_EMAIL"
            ],
            "x-enum-varnames": [
                "CodeValidationError",
                "CodeInvalidRequest",
                "CodeUnauthorized",
                "CodeForbidden",
                "CodeNotFound",
                "CodeConflict",
                "CodeDatabaseError",
                "CodeInternalError",
                "CodeInvalidToken",
                "CodeMissingToken",
                "CodeInvalidEmail"
            ]
        },
        "dto.ErrorData": {
            "type": "object",
            "properties": {
                "code": {
                    "$ref": "#/definitions/dto.ErrorCode"
                },
                "details": {
                    "type": "object",
                    "additionalProperties": {}
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "dto.ErrorEnvelope": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/dto.ErrorData"
                },
                "debug": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                },
                "trace_id": {
                    "type": "string"
                }
            }
        },
...
        "dto.MeEnvelope": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/dto.MeResponse"
                },
                "debug": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                },
                "trace_id": {
                    "type": "string"
                }
            }
        },
        "dto.MeResponse": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "firstname": {
                    "type": "string"
                },
                "lastname": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                },
                "user_type": {
                    "$ref": "#/definitions/models.UserType"
                }
            }
        },
...
        "models.UserType": {
            "type": "string",
            "enum": [
                "standard",
                "admin"
            ],
            "x-enum-varnames": [
                "StandardUser",
                "AdminUser"
            ]
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}